bot/requirements.txt
python-dotenv
selenium
requests
PyGithub
ollama-python
